server/requirements.txt
fastapi>=0.135,<0.136
uvicorn[standard]>=0.42,<0.43
python-jose[cryptography]>=3.5,<3.6
python-multipart>=0.0.22,<0.0.23
PyPDF2>=3.0.1,<3.1
scikit-learn>=1.8,<1.9
pandas>=3.0,<3.1
numpy>=2.4,<2.5
joblib>=1.5,<1.6
pydantic>=2.12,<2.13
bcrypt>=5.0,<5.1
nltk>=3.9,<4.0
